# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PaperTrade Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone number for verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /market/options:
    get:
      operationId: getOptionsChain
      tags: [market]
      summary: Get options chain for an index
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: expiry
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Options chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptionsChain"
  /market/expiries:
    get:
      operationId: getExpiries
      tags: [market]
      summary: Get available expiry dates for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of expiry dates
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /market/search:
    get:
      operationId: searchInstruments
      tags: [market]
      summary: Search instruments
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Instrument"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterInput:
      type: object
      required: [fullName, email, password, phone, age, city, otp]
      properties:
        fullName:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
        age:
          type: integer
          minimum: 18
        city:
          type: string
        otp:
          type: string
          minLength: 6
          maxLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        city:
          type: string
        stopLossLimit:
          type: ["number", "null"]
        targetPrice:
          type: ["number", "null"]
        maxDailyLoss:
          type: ["number", "null"]
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, fullName, email, phone, age, city, balance, initialCapital, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        age:
          type: integer
        city:
          type: string
        balance:
          type: number
        initialCapital:
          type: number
        stopLossLimit:
          type: ["number", "null"]
        targetPrice:
          type: ["number", "null"]
        maxDailyLoss:
          type: ["number", "null"]
        createdAt:
          type: string
...
    OptionsChain:
      type: object
      required: [symbol, expiry, underlyingPrice, options]
      properties:
        symbol:
          type: string
        expiry:
          type: string
        underlyingPrice:
          type: number
        options:
          type: array
          items:
            $ref: "#/components/schemas/OptionContract"
    OptionContract:
      type: object
      required: [strikePrice, type, premium, ltp, bidPrice, askPrice, lotSize, expiry, oi, volume]
      properties:
        strikePrice:
          type: number
        type:
          type: string
          enum: [CE, PE]
        premium:
          type: number
        ltp:
          type: number
        bidPrice:
          type: number
        askPrice:
          type: number
        lotSize:
          type: integer
        expiry:
          type: string
        oi:
          type: integer
        volume:
          type: integer
        iv:
          type: ["number", "null"]
        delta:
          type: ["number", "null"]
    Instrument:
      type: object
      required: [symbol, name, type]
      properties:
        symbol:
          type: string
        name:
          type: string
        type:
          type: string
...
    OtpRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    OtpResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        otp:
          type: string
          description: "Demo only: OTP returned in response since no SMS provider is configured"