# lib/api-zod/src/generated/types/otpResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PaperTrade Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OtpResponse {
  success: boolean;
  message: string;
  /** Demo only: OTP returned in response since no SMS provider is configured */
  otp?: string;
}