# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PaperTrade Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @minLength 3 */
  fullName: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone: string;
  /** @minimum 18 */
  age: number;
  city: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  otp: string;
}